+++ client/src/Components/SignComp/BackgroundImage.js
@@ -1,0 +1,28 @@
+import React from 'react';
+import { CssBaseline, Grid } from '@mui/material/';
+
+const BackgroundImage = () => {
+  return (
+    <>
+      <CssBaseline />
+      <Grid
+        item
+        xs={false}
+        sm={4}
+        md={7}
+        sx={{
+          backgroundImage: 'url(https://source.unsplash.com/random)',
+          backgroundRepeat: 'no-repeat',
+          backgroundColor: t =>
+            t.palette.mode === 'light'
+              ? t.palette.grey[50]
+              : t.palette.grey[900],
+          backgroundSize: 'cover',
+          backgroundPosition: 'center',
+        }}
+      />
+    </>
+  );
+};
+
+export default BackgroundImage;
