+++ client/src/Pages/Signup.js
@@ -1,134 +1,6 @@
 import * as React from 'react';
-import {
-  Avatar,
-  Button,
-  CssBaseline,
-  TextField,
-  FormControlLabel,
-  Checkbox,
-  Paper,
-  Grid,
-  Typography,
-  Box,
-} from '@mui/material/';
-import LockOutlinedIcon from '@mui/icons-material/LockOutlined';
-import { ThemeProvider } from '@mui/material/styles';
-import { palette, themeUserPage } from '../Styles/theme';
+import Sign from '../Components/SignComp/Sign';
 
 export default function Signup() {
-  const handleSubmit = event => {
-    event.preventDefault();
-    const data = new FormData(event.currentTarget);
-    console.log({
-      nickname: data.get('nickname'),
-      email: data.get('email'),
-      password: data.get('password'),
-      passwordCheck: data.get('passwordCheck'),
-    });
-  };
-
-  return (
-    <ThemeProvider theme={themeUserPage}>
-      <Grid container component="main" sx={{ height: '100vh' }}>
-        <CssBaseline />
-        <Grid
-          item
-          xs={false}
-          sm={4}
-          md={7}
-          sx={{
-            backgroundImage: 'url(https://source.unsplash.com/random)',
-            backgroundRepeat: 'no-repeat',
-            backgroundColor: t =>
-              t.palette.mode === 'light'
-                ? t.palette.grey[50]
-                : t.palette.grey[900],
-            backgroundSize: 'cover',
-            backgroundPosition: 'center',
-          }}
-        />
-        <Grid item xs={12} sm={8} md={5} component={Paper} elevation={6} square>
-          <Box
-            sx={{
-              my: 8,
-              mx: 4,
-              display: 'flex',
-              flexDirection: 'column',
-              alignItems: 'center',
-            }}
-          >
-            <Avatar sx={{ m: 1, bgcolor: palette.colorJamRealtime }}>
-              <LockOutlinedIcon />
-            </Avatar>
-            <Typography component="h1" variant="h5">
-              회원가입
-            </Typography>
-            <Box
-              component="form"
-              noValidate
-              onSubmit={handleSubmit}
-              sx={{ mt: 1 }}
-            >
-              <TextField
-                margin="normal"
-                required
-                fullWidth
-                id="nickname"
-                label="닉네임"
-                name="nickname"
-                autoComplete="nickname"
-                autoFocus
-                size="small"
-              />
-              <TextField
-                margin="normal"
-                required
-                fullWidth
-                id="email"
-                label="이메일"
-                name="email"
-                autoComplete="email"
-                size="small"
-              />
-              <TextField
-                margin="normal"
-                required
-                fullWidth
-                name="password"
-                label="비밀번호"
-                type="password"
-                id="password"
-                autoComplete="current-password"
-                size="small"
-              />
-              <TextField
-                margin="normal"
-                required
-                fullWidth
-                name="passwordCheck"
-                label="비밀번호 확인"
-                type="password"
-                id="passwordCheck"
-                autoComplete="current-password"
-                size="small"
-              />
-              <FormControlLabel
-                control={<Checkbox value="remember" color="primary" />}
-                label="회원가입 약관에 동의합니다"
-              />
-              <Button
-                color="primary"
-                type="submit"
-                fullWidth
-                variant="contained"
-                sx={{ mt: 3, mb: 2, boxShadow: 0 }}
-              >
-                회원가입
-              </Button>
-            </Box>
-          </Box>
-        </Grid>
-      </Grid>
-    </ThemeProvider>
-  );
+  return <Sign />;
 }
